refactor: white vs black box testing
