+++ internal/worker/pebblelokiconfig/agent_mock_test.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -package pebblelokiconfig_test -destination agent_mock_test.go github.com/juju/juju/agent Agent,Config
+//	mockgen -package pebblelokiconfig -destination agent_mock_test.go github.com/juju/juju/agent Agent,Config
 //
 
-// Package pebblelokiconfig_test is a generated GoMock package.
-package pebblelokiconfig_test
+// Package pebblelokiconfig is a generated GoMock package.
+package pebblelokiconfig
 
 import (
 	time "time"
