+++ internal/worker/pebblelokiconfig/clock_mock_test.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -package pebblelokiconfig_test -destination clock_mock_test.go github.com/juju/clock Clock
+//	mockgen -package pebblelokiconfig -destination clock_mock_test.go github.com/juju/clock Clock
 //
 
-// Package pebblelokiconfig_test is a generated GoMock package.
-package pebblelokiconfig_test
+// Package pebblelokiconfig is a generated GoMock package.
+package pebblelokiconfig
 
 import (
 	time "time"
