+++ internal/worker/pebblelokiconfig/notify_watcher_mock_test.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -package pebblelokiconfig_test -destination notify_watcher_mock_test.go github.com/juju/juju/core/watcher NotifyWatcher
+//	mockgen -package pebblelokiconfig -destination notify_watcher_mock_test.go github.com/juju/juju/core/watcher NotifyWatcher
 //
 
-// Package pebblelokiconfig_test is a generated GoMock package.
-package pebblelokiconfig_test
+// Package pebblelokiconfig is a generated GoMock package.
+package pebblelokiconfig
 
 import (
 	gomock "github.com/canonical/gomock/gomock"
