+++ internal/worker/pebblelokiconfig/pebble_client_mock_test.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -package pebblelokiconfig_test -destination pebble_client_mock_test.go github.com/juju/juju/internal/worker/pebblelokiconfig PebbleClient,LoggerAPI
+//	mockgen -package pebblelokiconfig -destination pebble_client_mock_test.go github.com/juju/juju/internal/worker/pebblelokiconfig PebbleClient,LoggerAPI
 //
 
-// Package pebblelokiconfig_test is a generated GoMock package.
-package pebblelokiconfig_test
+// Package pebblelokiconfig is a generated GoMock package.
+package pebblelokiconfig
 
 import (
 	context "context"
